+++ src/fastergerman/game/__init__.py
@@ -1,4 +1,4 @@
-from fastergerman.game.game import Game, Question, Score, Settings
+from fastergerman.game.game import Game, Question, Score, Settings, LanguageLevel
 from fastergerman.game.questions import FileQuestionsSource
 from fastergerman.game.game_store import AbstractGameStore, FileGameStore, InMemoryGameStore
 from fastergerman.game.game_timer import AbstractGameTimer, GameTimerState, GameTimer, TimerError